refactor: Spice86 compile-time DI class
